Unify cattr and mattr accessors declarations
