+++ activesupport/lib/active_support/core_ext/class.rb
@@ -1,4 +1,3 @@
 require 'active_support/core_ext/class/attribute'
-require 'active_support/core_ext/class/attribute_accessors'
 require 'active_support/core_ext/class/delegating_attributes'
 require 'active_support/core_ext/class/subclasses'
